requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-multipart==0.0.20
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-intel==2.17.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
